[Key Vault] Separating public and internal folders
