+++ sdk/keyvault/keyvault-certificates/test/internal/challengeBasedAuthenticationPolicy.spc.ts
@@ -2,17 +2,18 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import { CertificateClient } from "../src";
+import { createSandbox } from "sinon";
 import { env, Recorder } from "@azure/test-utils-recorder";
-import { authenticate } from "./utils/testAuthentication";
-import TestClient from "./utils/testClient";
+
+import { CertificateClient } from "../../src";
 import {
   AuthenticationChallengeCache,
   AuthenticationChallenge,
   parseWWWAuthenticate
-} from "../src/core/challengeBasedAuthenticationPolicy";
-import { createSandbox } from "sinon";
-import { testPollerProperties } from './utils/recorderUtils';
+} from "../../src/core/challengeBasedAuthenticationPolicy";
+import { testPollerProperties } from '../utils/recorderUtils';
+import { authenticate } from "../utils/testAuthentication";
+import TestClient from "../utils/testClient";
 
 // Following the philosophy of not testing the insides if we can test the outsides...
 // I present you with this "Get Out of Jail Free" card (in reference to Monopoly).
