+++ sdk/keyvault/keyvault-certificates/test/public/lro.create.spec.ts
@@ -2,14 +2,15 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import { CertificateClient, DeletedCertificate, DefaultCertificatePolicy } from "../src";
-import { testPollerProperties } from "./utils/recorderUtils";
-import { env, Recorder } from "@azure/test-utils-recorder";
-import { authenticate } from "./utils/testAuthentication";
-import TestClient from "./utils/testClient";
 import { PollerStoppedError } from "@azure/core-lro";
+import { env, Recorder } from "@azure/test-utils-recorder";
 
-describe("Certificates client - lro - delete", () => {
+import { CertificateClient, KeyVaultCertificate, DefaultCertificatePolicy } from "../../src";
+import { testPollerProperties } from "../utils/recorderUtils";
+import { authenticate } from "../utils/testAuthentication";
+import TestClient from "../utils/testClient";
+
+describe("Certificates client - LRO - create", () => {
   const certificatePrefix = `recover${env.CERTIFICATE_NAME || "CertificateName"}`;
   let certificateSuffix: string;
   let client: CertificateClient;
@@ -30,45 +31,39 @@ describe("Certificates client - lro - delete", () => {
 
   // The tests follow
 
-  it("can wait until a certificate is deleted", async function() {
+  it("can wait until a certificate is created", async function() {
     const certificateName = testClient.formatName(
       `${certificatePrefix}-${this!.test!.title}-${certificateSuffix}`
     );
-    await client.beginCreateCertificate(
+    const poller = await client.beginCreateCertificate(
       certificateName,
       DefaultCertificatePolicy,
       testPollerProperties
     );
-    const poller = await client.beginDeleteCertificate(certificateName, testPollerProperties);
     assert.ok(poller.getOperationState().isStarted);
 
-    // The pending deleted certificate can be obtained this way:
+    // The pending certificate can be obtained this way:
     assert.equal(poller.getOperationState().result!.name, certificateName);
 
-    let deletedCertificate: DeletedCertificate = await poller.pollUntilDone();
-    assert.equal(deletedCertificate.name, certificateName);
+    const createdCertificate: KeyVaultCertificate = await poller.pollUntilDone();
+    assert.equal(createdCertificate.name, certificateName);
     assert.ok(poller.getOperationState().isCompleted);
 
-    // Retrieving it without the poller
-    deletedCertificate = await client.getDeletedCertificate(certificateName);
-    assert.equal(deletedCertificate.name, certificateName);
-
-    // The final deleted certificate can also be obtained this way:
+    // The final certificate can also be obtained this way:
     assert.equal(poller.getOperationState().result!.name, certificateName);
 
-    await testClient.purgeCertificate(certificateName);
+    await testClient.flushCertificate(certificateName);
   });
 
   it("can resume from a stopped poller", async function() {
     const certificateName = testClient.formatName(
       `${certificatePrefix}-${this!.test!.title}-${certificateSuffix}`
     );
-    await client.beginCreateCertificate(
+    const poller = await client.beginCreateCertificate(
       certificateName,
       DefaultCertificatePolicy,
       testPollerProperties
     );
-    const poller = await client.beginDeleteCertificate(certificateName, testPollerProperties);
     assert.ok(poller.getOperationState().isStarted);
 
     poller.pollUntilDone().catch((e) => {
@@ -83,20 +78,20 @@ describe("Certificates client - lro - delete", () => {
 
     const serialized = poller.toString();
 
-    const resumePoller = await client.beginDeleteCertificate(certificateName, {
-      resumeFrom: serialized,
-      ...testPollerProperties
-    });
+    const resumePoller = await client.beginCreateCertificate(
+      certificateName,
+      DefaultCertificatePolicy,
+      {
+        resumeFrom: serialized,
+        ...testPollerProperties
+      }
+    );
 
     assert.ok(resumePoller.getOperationState().isStarted);
+    const createdCertificate: KeyVaultCertificate = await resumePoller.pollUntilDone();
+    assert.equal(createdCertificate.name, certificateName);
+    assert.ok(resumePoller.getOperationState().isCompleted);
 
-    let deletedCertificate: DeletedCertificate = await resumePoller.pollUntilDone();
-    assert.equal(deletedCertificate.name, certificateName);
-
-    // Retrieving it without the poller
-    deletedCertificate = await client.getDeletedCertificate(certificateName);
-    assert.equal(deletedCertificate.name, certificateName);
-
-    await testClient.purgeCertificate(certificateName);
+    await testClient.flushCertificate(certificateName);
   });
 });
