+++ sdk/keyvault/keyvault-certificates/test/public/lro.recover.spec.ts
@@ -2,13 +2,14 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import { CertificateClient, DeletedCertificate, DefaultCertificatePolicy } from "../src";
-import { testPollerProperties } from "./utils/recorderUtils";
 import { env, Recorder } from "@azure/test-utils-recorder";
-import { authenticate } from "./utils/testAuthentication";
-import TestClient from "./utils/testClient";
 import { PollerStoppedError } from "@azure/core-lro";
-import { assertThrowsAbortError } from "./utils/utils.common";
+
+import { CertificateClient, DeletedCertificate, DefaultCertificatePolicy } from "../../src";
+import { assertThrowsAbortError } from "../utils/utils.common";
+import { testPollerProperties } from "../utils/recorderUtils";
+import { authenticate } from "../utils/testAuthentication";
+import TestClient from "../utils/testClient";
 
 describe("Certificates client - LRO - recoverDelete", () => {
   const certificatePrefix = `recover${env.CERTIFICATE_NAME || "CertificateName"}`;
