+++ sdk/keyvault/keyvault-keys/test/public/lro.recoverDelete.spec.ts
@@ -2,14 +2,16 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import { KeyClient, DeletedKey } from "../src";
-import { testPollerProperties } from "./utils/recorderUtils";
 import { env, Recorder } from "@azure/test-utils-recorder";
-import { authenticate } from "./utils/testAuthentication";
-import TestClient from "./utils/testClient";
 import { PollerStoppedError } from "@azure/core-lro";
 
-describe("Keys client - Long Running Operations - delete", () => {
+import { KeyClient, DeletedKey } from "../../src";
+import { assertThrowsAbortError } from "../utils/utils.common";
+import { testPollerProperties } from "../utils/recorderUtils";
+import { authenticate } from "../utils/testAuthentication";
+import TestClient from "../utils/testClient";
+
+describe("Keys client - Long Running Operations - recoverDelete", () => {
   const keyPrefix = `recover${env.CERTIFICATE_NAME || "KeyName"}`;
   let keySuffix: string;
   let client: KeyClient;
@@ -30,13 +32,17 @@ describe("Keys client - Long Running Operations - delete", () => {
 
   // The tests follow
 
-  it("can wait until a key is deleted", async function() {
+  it("can wait until a key is recovered", async function() {
     const keyName = testClient.formatName(`${keyPrefix}-${this!.test!.title}-${keySuffix}`);
     await client.createKey(keyName, "RSA");
-    const poller = await client.beginDeleteKey(keyName, testPollerProperties);
+
+    const deletePoller = await client.beginDeleteKey(keyName, testPollerProperties);
+    await deletePoller.pollUntilDone();
+
+    const poller = await client.beginRecoverDeletedKey(keyName, testPollerProperties);
     assert.ok(poller.getOperationState().isStarted);
 
-    // The pending deleted can be obtained this way:
+    // The pending key can be obtained this way:
     assert.equal(poller.getOperationState().result!.name, keyName);
 
     const deletedKey: DeletedKey = await poller.pollUntilDone();
@@ -46,13 +52,16 @@ describe("Keys client - Long Running Operations - delete", () => {
     // The final key can also be obtained this way:
     assert.equal(poller.getOperationState().result!.name, keyName);
 
-    await testClient.purgeKey(keyName);
+    await testClient.flushKey(keyName);
   });
 
   it("can resume from a stopped poller", async function() {
     const keyName = testClient.formatName(`${keyPrefix}-${this!.test!.title}-${keySuffix}`);
     await client.createKey(keyName, "RSA");
-    const poller = await client.beginDeleteKey(keyName, testPollerProperties);
+    const deletePoller = await client.beginDeleteKey(keyName, testPollerProperties);
+    await deletePoller.pollUntilDone();
+
+    const poller = await client.beginRecoverDeletedKey(keyName, testPollerProperties);
     assert.ok(poller.getOperationState().isStarted);
 
     poller.pollUntilDone().catch((e) => {
@@ -69,16 +78,31 @@ describe("Keys client - Long Running Operations - delete", () => {
 
     const serialized = poller.toString();
 
-    const resumePoller = await client.beginDeleteKey(keyName, {
+    const resumePoller = await client.beginRecoverDeletedKey(keyName, {
       resumeFrom: serialized,
       ...testPollerProperties
     });
 
-    assert.ok(resumePoller.getOperationState().isStarted);
+    assert.ok(poller.getOperationState().isStarted);
     const deletedKey: DeletedKey = await resumePoller.pollUntilDone();
     assert.equal(deletedKey.name, keyName);
     assert.ok(resumePoller.getOperationState().isCompleted);
 
-    await testClient.purgeKey(keyName);
+    await testClient.flushKey(keyName);
+  });
+
+  // On playback mode, the tests happen too fast for the timeout to work
+  it("can recover a deleted key with requestOptions timeout", async function() {
+    recorder.skip(undefined, "Timeout tests don't work on playback mode.");
+    const keyName = testClient.formatName(`${keyPrefix}-${this!.test!.title}-${keySuffix}`);
+    await client.createKey(keyName, "RSA");
+    const deletePoller = await client.beginDeleteKey(keyName, testPollerProperties);
+    await deletePoller.pollUntilDone();
+    await assertThrowsAbortError(async () => {
+      await client.beginRecoverDeletedKey(keyName, {
+        requestOptions: { timeout: 1 },
+        ...testPollerProperties
+      });
+    });
   });
 });
