+++ sdk/keyvault/keyvault-secrets/test/public/CRUD.spec.ts
@@ -2,13 +2,14 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import { SecretClient } from "../../src";
-import { testPollerProperties } from "../utils/recorderUtils";
 import { env, Recorder } from "@azure/test-utils-recorder";
+import { AbortController } from "@azure/abort-controller";
+
+import { SecretClient } from "../../src";
+import { assertThrowsAbortError } from "../utils/utils.common";
+import { testPollerProperties } from "../utils/recorderUtils";
 import { authenticate } from "../utils/testAuthentication";
 import TestClient from "../utils/testClient";
-import { AbortController } from "@azure/abort-controller";
-import { assertThrowsAbortError } from "../utils/utils.common";
 
 describe("Secret client - create, read, update and delete operations", () => {
   const secretValue = "SECRET_VALUE";
