+++ sdk/keyvault/keyvault-secrets/test/public/list.spec.ts
@@ -3,12 +3,14 @@
 
 import * as assert from "assert";
 import chai from "chai";
-import { SecretClient } from "../../src";
-import { testPollerProperties } from "../utils/recorderUtils";
 import { env, Recorder, isRecordMode } from "@azure/test-utils-recorder";
+
+import { SecretClient } from "../../src";
+import { assertThrowsAbortError } from "../utils/utils.common";
+import { testPollerProperties } from "../utils/recorderUtils";
 import { authenticate } from "../utils/testAuthentication";
 import TestClient from "../utils/testClient";
-import { assertThrowsAbortError } from "../utils/utils.common";
+
 const { expect } = chai;
 
 describe("Secret client - list secrets in various ways", () => {
