+++ sdk/keyvault/keyvault-secrets/test/public/recoverBackupRestore.spec.ts
@@ -2,13 +2,14 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import { SecretClient } from "../../src";
 import { isNode } from "@azure/core-http";
-import { testPollerProperties } from "../utils/recorderUtils";
 import { env, isPlaybackMode, Recorder, delay, isRecordMode } from "@azure/test-utils-recorder";
+
+import { SecretClient } from "../../src";
+import { assertThrowsAbortError } from "../utils/utils.common";
+import { testPollerProperties } from "../utils/recorderUtils";
 import { authenticate } from "../utils/testAuthentication";
 import TestClient from "../utils/testClient";
-import { assertThrowsAbortError } from "../utils/utils.common";
 
 describe("Secret client - restore secrets and recover backups", () => {
   const secretPrefix = `CRUD${env.SECRET_NAME || "SecretName"}`;
